+++ lib/zenlish/wclasses/adverb_there.rb
@@ -1,0 +1,9 @@
+require_relative 'adverb'
+
+module Zenlish
+  module WClasses
+    # TODO: document
+    class AdverbThere < Adverb
+    end # class
+  end # module
+end # module