+++ lib/zenlish/wclasses/auxiliary.rb
@@ -1,0 +1,11 @@
+require_relative 'verb'
+
+module Zenlish
+  module WClasses
+    # The verbs: be, have, do
+    # As they are so important to the English language, they deserve
+    # their own class
+    class Auxiliary < Verb
+    end # class
+  end # module
+end # module