+++ lib/zenlish/wclasses/auxiliary_be.rb
@@ -1,0 +1,9 @@
+require_relative 'auxiliary'
+
+module Zenlish
+  module WClasses
+    # TODO: document
+    class AuxiliaryBe < Auxiliary
+    end # class
+  end # module
+end # module