+++ lib/zenlish/wclasses/irregular_verb_be.rb
@@ -1,0 +1,9 @@
+require_relative 'irregular_verb'
+
+module Zenlish
+  module WClasses
+    # TODO: document
+    class IrregularVerbBe < IrregularVerb
+    end # class
+  end # module
+end # module