+++ lib/zenlish/wclasses/preposition_than.rb
@@ -1,0 +1,9 @@
+require_relative 'preposition_than'
+
+module Zenlish
+  module WClasses
+    # the 'than' preposition
+    class PrepositionThan < Preposition
+    end # class
+  end # module
+end # module